feat: Adicionando a remoção dos logs no encerramento do sistema

+++ Ocorrencias/abertura_urna.py
@@ -67,4 +67,5 @@ def excluir_ocorrencia_abertura_urna():
     Returns: None
     """
     arquivo = CAMINHO_ARQUIVO
-    os.remove(arquivo)
+    if os.path.exists(arquivo):
+        os.remove(arquivo)
