+++ Ocorrencias/encerramento_urna.py
@@ -98,5 +98,5 @@ def excluir_ocorrencia_encerramento_urna():
     """
 
     arquivo = CAMINHO_ARQUIVO
-
-    os.remove(arquivo)
+    if os.path.exists(arquivo):
+        os.remove(arquivo)
