+++ Ocorrencias/voto_computado.py
@@ -71,4 +71,5 @@ def excluir_arquivo_voto_computado():
             None
     """
     arquivo = CAMINHO_ARQUIVO
-    os.remove(arquivo)
+    if os.path.exists(arquivo):
+        os.remove(arquivo)
